+++ Tortillapp-web/Control/UserController.cs
@@ -1,7 +1,4 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore.Metadata.Internal;
-using System.Web;
-using Tortillapp_web.Data;
 using Tortillapp_web.Model;
 
 namespace Tortillapp_web.Control
@@ -13,34 +10,10 @@ namespace Tortillapp_web.Control
             return View();
         }
 
-        [HttpGet]
-        public ActionResult Login()
+        public IActionResult Index(UserData u)
         {
-            return View();
-        }
-
-        [HttpPost]
-        public ActionResult Login(Model.UserData user)
-        {
-            if(ModelState.IsValid)
-            {
-                if(user.UserMail.Equals(user.UserMail))
-                {
-                    if(user.UserPass.Equals(user.UserPass))
-                    {
-                        return RedirectToAction("Index");
-                    }
-                    else
-                    {
-                        ModelState.AddModelError("", "La contraseña o el correo son incorrectos");
-                    }
-                }
-            }
-            return View(user);
-        }
-        public ActionResult Logout() 
-        {
-            return RedirectToAction("Index");
+            return View(u);
         }
     }
 }
+
