+++ Tortillapp-web/Pages/Recipe/_Create.cshtml
@@ -1,0 +1,63 @@
+﻿@page
+@model Tortillapp_web.Pages.Receta.CreateModel
+
+@{
+    ViewData["Title"] = "Create";
+}
+
+<h1>Create</h1>
+
+<h4>RecipeInfo</h4>
+<hr />
+<div class="row">
+    <div class="col-md-4">
+        <form method="post">
+            <div asp-validation-summary="ModelOnly" class="text-danger"></div>
+            <div class="form-group">
+                <label asp-for="RecipeInfo.UserId" class="control-label"></label>
+                <select asp-for="RecipeInfo.UserId" class ="form-control" asp-items="ViewBag.UserId"></select>
+            </div>
+            <div class="form-group">
+                <label asp-for="RecipeInfo.RecipeTitle" class="control-label"></label>
+                <input asp-for="RecipeInfo.RecipeTitle" class="form-control" />
+                <span asp-validation-for="RecipeInfo.RecipeTitle" class="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <label asp-for="RecipeInfo.Published" class="control-label"></label>
+                <input asp-for="RecipeInfo.Published" class="form-control" />
+                <span asp-validation-for="RecipeInfo.Published" class="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <label asp-for="RecipeInfo.RecipePortion" class="control-label"></label>
+                <input asp-for="RecipeInfo.RecipePortion" class="form-control" />
+                <span asp-validation-for="RecipeInfo.RecipePortion" class="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <label asp-for="RecipeInfo.RecipeTime" class="control-label"></label>
+                <input asp-for="RecipeInfo.RecipeTime" class="form-control" />
+                <span asp-validation-for="RecipeInfo.RecipeTime" class="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <label asp-for="RecipeInfo.RecipePic" class="control-label"></label>
+                <input asp-for="RecipeInfo.RecipePic" class="form-control" />
+                <span asp-validation-for="RecipeInfo.RecipePic" class="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <label asp-for="RecipeInfo.RecipeTips" class="control-label"></label>
+                <input asp-for="RecipeInfo.RecipeTips" class="form-control" />
+                <span asp-validation-for="RecipeInfo.RecipeTips" class="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <input type="submit" value="Create" class="btn btn-primary" />
+            </div>
+        </form>
+    </div>
+</div>
+
+<div>
+    <a asp-page="Index">Back to List</a>
+</div>
+
+@section Scripts {
+    @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
+}
