+++ Tortillapp-web/Pages/Shared/_Layout.cshtml
@@ -79,7 +79,7 @@ else { isLogged = 0; }
                                     <ul class="dropdown-menu" aria-labelledby="navbarDropdownLink">
                                         <li><a class="dropdown-item" asp-page="/MyProfile" asp-page-handler="@actualUser">Mi perfil</a></li>
                                         <li><a class="dropdown-item" asp-page="#">Recetas favoritas</a></li>
-                                        <li><a class="dropdown-item" asp-page="/Recipe/NewRecipe">Crear receta</a></li>
+                                        <li><a class="dropdown-item" asp-page="/Recipe/Index">Crear receta</a></li>
                                         <li><a class="dropdown-item" asp-page="#">Acerca de</a></li>
                                         <li><a class="dropdown-item" asp-page="/Index" asp-page-handler="Logout">Cerrar sesión</a></li>
                                     </ul>
