+++ Tortillapp-web/Program.cs
@@ -17,7 +17,7 @@ builder.Services.AddDistributedMemoryCache();
 
 builder.Services.AddSession(options =>
 {
-    options.IdleTimeout = TimeSpan.FromSeconds(60);
+    options.IdleTimeout = TimeSpan.FromSeconds(600);
     options.Cookie.HttpOnly = true;
 });
